add script that generates kava genesis.json
the script generates a genesis file from `kava init` and iteratively
updates it to a desired starting state for kvtool's master kava template.

the only change to the genesis in this commit _should_ be that the
validator address has been updated to match the validator listed in the
addresses file. all other existing accounts and positions should be
named and initially funded just as before. all modules have been checked
to match the original genesis.