fix: Ensure chain upgrades when specified (#203)
Retries upgrade proposal vote, as cosmos sdk v0.47.0 removed broadcast mode block. With the "sync" mode, it only runs CheckTx and may not be in a block in subsequent calls (causes account number mismatch error).

Ensure chain actually halted by monitoring chain logs. This was a source of vague errors when the upgrade proposal was not correctly applied then continued to make blocks. The binary replacement would still continue and cause errors due to a missing upgrade.

Note: Ideally we replace the use of exec.Command() for docker with the docker go client

Add evmutil balances to genesis for testing purposes.