fix: Update generate genesis to work on linux (#204)
On darwin, sed is the BSD version which requires -i to have an argument
and either a space, or an equal sign for the flag argument.  It does not
allow no argument.

On linux, sed is usually the GNU version, which allows -i to optionally
have an argument and either have no space or an equal sign.

For example:
  BSD - `sed -i ''` and `sed -i=''` is valid
  GNU - `sed -i`, `sed -i''` and `sed -i=''` is valid

Therefore, we update the genesis script to use the latter and enable
better cross platform support.