feat: Add valid precompile state to genesis (#206)
This adds 0x9000000000000000000000000000000000000001 to state as an
enabled precompiled contract to the EVM parameters.  This also requires
us to ensure an EthAccount is created that matches the 0x address, the
EthAccount has code hash equal to keccak256(0x01), a sequence of 1, and
a nil public key.  In addition, we must create an evm genesis account
with code equal to 0x01 and optional storage.  This extends
eveniy-scherbina's work on adding enabled precompiles to state.

Co-authored-by: evgeniy-scherbina <evgeniy.shcherbina.es@gmail.com>